Add tool for retrieving television show data
This PR includes restructing of some entities to create a
generic media discovery infrastructure (shared code) on top
of which tools are added to fetch specific types of media.

At this time, the tooling supports two types of media: film
(existing functionality) and television shows (new functionality).
In the future, I hope to add tooling for more content types.